pyproject.toml
[project]
name = "polymarket-monitor-engine"
version = "0.1.0"
description = "Polymarket finance/geopolitics monitor with fan-out sinks."
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [
  { name = "Polymarket Monitor" }
]
dependencies = [
  "httpx>=0.27",
  "websockets>=12.0",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "pyyaml>=6.0",
  "structlog>=24.1",
  "redis>=5.0",
  "orjson>=3.10",
  "tenacity>=9.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
  "mypy>=1.10",
  "types-redis>=4.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
src = ["src", "tests"]
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "ASYNC", "N"]
ignore = ["E203", "E266"]

[tool.ruff.lint.isort]
known-first-party = ["polymarket_monitor_engine"]

[tool.ruff.format]
line-ending = "lf"

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.14"
strict = true
warn_unused_configs = true

